Separate out DataTreeEvaluator class in a different file. Update action fields for dynamic binding path list
