+++ app/client/src/workers/evaluationUtils.ts
@@ -2,10 +2,10 @@ import {
   DependencyMap,
   isChildPropertyPath,
   isDynamicValue,
-} from "../utils/DynamicBindingUtils";
-import { WidgetType } from "../constants/WidgetConstants";
-import { WidgetProps } from "../widgets/BaseWidget";
-import { WidgetTypeConfigMap } from "../utils/WidgetFactory";
+} from "utils/DynamicBindingUtils";
+import { WidgetType } from "constants/WidgetConstants";
+import { WidgetProps } from "widgets/BaseWidget";
+import { WidgetTypeConfigMap } from "utils/WidgetFactory";
 import { VALIDATORS } from "./validations";
 import { Diff } from "deep-diff";
 import {
@@ -14,7 +14,7 @@ import {
   DataTreeEntity,
   DataTreeWidget,
   ENTITY_TYPE,
-} from "../entities/DataTree/dataTreeFactory";
+} from "entities/DataTree/dataTreeFactory";
 import _ from "lodash";
 
 // Dropdown1.options[1].value -> Dropdown1.options[1]
@@ -367,4 +367,100 @@ export const trimDependantChangePaths = (
     }
   }
   return trimmedPaths;
+};
+
+export const addFunctions = (dataTree: Readonly<DataTree>): DataTree => {
+  const withFunction: DataTree = _.cloneDeep(dataTree);
+  withFunction.actionPaths = [];
+  Object.keys(withFunction).forEach((entityName) => {
+    const entity = withFunction[entityName];
+    if (isAction(entity)) {
+      const runFunction = function(
+        this: DataTreeAction,
+        onSuccess: string,
+        onError: string,
+        params = "",
+      ) {
+        return {
+          type: "RUN_ACTION",
+          payload: {
+            actionId: this.actionId,
+            onSuccess: onSuccess ? `{{${onSuccess.toString()}}}` : "",
+            onError: onError ? `{{${onError.toString()}}}` : "",
+            params,
+          },
+        };
+      };
+      _.set(withFunction, `${entityName}.run`, runFunction);
+      withFunction.actionPaths &&
+        withFunction.actionPaths.push(`${entityName}.run`);
+    }
+  });
+  withFunction.navigateTo = function(
+    pageNameOrUrl: string,
+    params: Record<string, string>,
+    target?: string,
+  ) {
+    return {
+      type: "NAVIGATE_TO",
+      payload: { pageNameOrUrl, params, target },
+    };
+  };
+  withFunction.actionPaths.push("navigateTo");
+
+  withFunction.showAlert = function(message: string, style: string) {
+    return {
+      type: "SHOW_ALERT",
+      payload: { message, style },
+    };
+  };
+  withFunction.actionPaths.push("showAlert");
+
+  withFunction.showModal = function(modalName: string) {
+    return {
+      type: "SHOW_MODAL_BY_NAME",
+      payload: { modalName },
+    };
+  };
+  withFunction.actionPaths.push("showModal");
+
+  withFunction.closeModal = function(modalName: string) {
+    return {
+      type: "CLOSE_MODAL",
+      payload: { modalName },
+    };
+  };
+  withFunction.actionPaths.push("closeModal");
+
+  withFunction.storeValue = function(key: string, value: string) {
+    return {
+      type: "STORE_VALUE",
+      payload: { key, value },
+    };
+  };
+  withFunction.actionPaths.push("storeValue");
+
+  withFunction.download = function(data: string, name: string, type: string) {
+    return {
+      type: "DOWNLOAD",
+      payload: { data, name, type },
+    };
+  };
+  withFunction.actionPaths.push("download");
+
+  withFunction.copyToClipboard = function(
+    data: string,
+    options?: { debug?: boolean; format?: string },
+  ) {
+    return {
+      type: "COPY_TO_CLIPBOARD",
+      payload: {
+        data,
+        options: { debug: options?.debug, format: options?.format },
+      },
+    };
+  };
+  withFunction.actionPaths.push("copyToClipboard");
+
+  return withFunction;
 };
